+++ web/src/state_data.ts
@@ -161,6 +161,7 @@ export const custom_profile_field_schema = z.object({
     order: z.number(),
     required: z.boolean(),
     type: z.number(),
+    use_for_user_matching: z.optional(z.boolean()),
 });
 
 export type CustomProfileField = z.output<typeof custom_profile_field_schema>;